navigation fixes and refactoring
