+++ src/components/HomePage/HomePage.jsx
@@ -2,7 +2,7 @@
  * This code was generated by Builder.io
  */
 import React from "react";
-import Header from "./Header";
+import Header from "../common/Header";
 import Hero from "./Hero";
 import TrustedBy from "./TrustedBy";
 import ComplianceRequirements from "./ComplianceRequirements";
@@ -11,10 +11,10 @@ import OneCentralPlatform from "./OneCentralPlatform";
 import CaseStudies from "./CaseStudies";
 import HowItWorks from "./HowItWorks";
 import AutomateReporting from "./AutomateReporting";
-import FAQ from "./FAQ";
-import ContactUs from "./ContactUs";
-import GetStarted from "./GetStarted";
-import Footer from "./Footer";
+import FAQ from "../common/FAQ";
+import ContactUs from "../common/ContactUs";
+import GetStarted from "../common/GetStarted";
+import Footer from "../common/Footer";
 
 function HomePage() {
   return (
