+++ src/components/RimbaComponents/HowIt.jsx
@@ -2,16 +2,16 @@
  * This code was generated by Builder.io
  */
 import React from "react";
-import Header from "./Header";
 import MainContent from "./MainContent";
 // import Footer from "./Footer";
-import GetStarted from "../HomePage/GetStarted";
-import Footer from "../HomePage/Footer";
+import GetStarted from "../common/GetStarted";
+import Footer from "../common/Footer";
+import Header from "../common/Header";
 
 const HowIt = () => {
   return (
     <div className="flex overflow-hidden flex-col items-center bg-white">
-      <Header />
+      <Header></Header>
       <MainContent />
       <GetStarted/>
       <Footer />
