+++ src/components/aboutUs/AboutUs.jsx
@@ -1,25 +1,15 @@
-/**
- * This code was generated by Builder.io
- */
 import React from "react";
-import Navbar from "./Navbar";
 import AboutHero from "./AboutHero";
-// import Features from "./Features";
 import Stats from "./Stats";
-// import FAQ from "./FAQ";
-// import ContactSection from "./ContactSection";
-// import GetStarted from "./GetStarted";
-// import Footer from "./Footer";
-
-import GetStarted from "../HomePage/GetStarted";
-import Footer from "../HomePage/Footer";
-import ContactUs from "../HomePage/ContactUs";
-import FAQ from "../HomePage/FAQ";
-
+import GetStarted from "../common/GetStarted";
+import Footer from "../common/Footer";
+import ContactUs from "../common/ContactUs";
+import FAQ from "../common/FAQ";
+import Header from "../common/Header";
 function AboutUs() {
   return (
     <main className="flex overflow-hidden flex-col items-center bg-white">
-      <Navbar />
+      <Header></Header>
       <AboutHero />
       <Stats />
       <FAQ></FAQ>
