--- src/components/aboutUs/Navbar.jsx
@@ -1,43 +1,0 @@
-/**
- * This code was generated by Builder.io
- */
-import React from "react";
-
-const navItems = [
-  { label: "Home", href: "#" },
-  { label: "How it Works", href: "#" },
-  { label: "What's New", href: "#" },
-  { label: "Case Studies", href: "#" },
-  { label: "About Us", href: "#", active: true },
-];
-
-function Navbar() {
-  return (
-    <header className="flex flex-wrap gap-5 justify-between self-stretch px-20 py-4 w-full text-center bg-white bg-opacity-40 shadow-[0px_4px_20px_rgba(209,234,232,0.3)] max-md:px-5 max-md:max-w-full">
-      <img
-        loading="lazy"
-        src="https://cdn.builder.io/api/v1/image/assets/TEMP/8e49f884a54337781454ce8d21c6fa0c281e0e6c5e3ca6cc2562ffe4f58afdb9?placeholderIfAbsent=true&apiKey=fd087a8e03ab4599addc57757630651c"
-        alt="Rimba Logo"
-        className="object-contain shrink-0 self-start max-w-full aspect-[3.04] w-[140px]"
-      />
-      <nav className="flex flex-wrap gap-8 justify-center items-center my-auto text-base text-slate-950 text-opacity-60 max-md:max-w-full">
-        {navItems.map((item, index) => (
-          <a
-            key={index}
-            href={item.href}
-            className={`self-stretch my-auto ${
-              item.active ? "font-bold text-teal-700" : ""
-            }`}
-          >
-            {item.label}
-          </a>
-        ))}
-      </nav>
-      <button className="gap-2.5 self-stretch py-4 px-6 text-sm font-medium text-white bg-teal-700 rounded-xl min-h-[48px] max-md:px-5">
-        Request A Demo
-      </button>
-    </header>
-  );
-}
-
-export default Navbar;
