+++ src/components/caseStudies/CaseStudiesFrame.jsx
@@ -1,10 +1,10 @@
 import React from "react";
-import Header from "../HomePage/Header";
+import Header from "../common/Header";
 import CaseStudies from "./CaseStudies";
-import GetStarted from "../HomePage/GetStarted";
-import Footer from "../HomePage/Footer";
-import ContactUs from "../HomePage/ContactUs";
-import FAQ from "../HomePage/FAQ";
+import GetStarted from "../common/GetStarted";
+import Footer from "../common/Footer";
+import ContactUs from "../common/ContactUs";
+import FAQ from "../common/FAQ";
 function CaseStudiesFrame(){
     return(
     <div className="flex overflow-hidden flex-col items-center bg-white">
