+++ src/components/common/FAQ.jsx
