+++ src/components/common/Header.jsx
@@ -1,9 +1,11 @@
-/**
- * This code was generated by Builder.io
- */
 import React from "react";
+import { useLocation } from "react-router-dom";
 
 function Header() {
+  const location = useLocation();
+
+  const isActive = (path) => location.pathname === path;
+
   return (
     <header className="flex flex-wrap gap-5 justify-between self-stretch px-20 py-4 w-full bg-white bg-opacity-40 shadow-[0px_4px_20px_rgba(209,234,232,0.3)] max-md:px-5 max-md:max-w-full">
       <img
@@ -13,19 +15,44 @@ function Header() {
         alt="Company logo"
       />
       <nav className="flex flex-wrap gap-8 justify-center items-center my-auto text-base text-slate-950 text-opacity-60 max-md:max-w-full">
-        <a href="/" className="self-stretch my-auto font-bold text-teal-700">
+        <a
+          href="/"
+          className={`self-stretch my-auto ${
+            isActive("/") ? "text-teal-700" : "text-slate-950"
+          }`}
+        >
           Home
         </a>
-        <a href="/how-it-works" className="self-stretch my-auto">
+        <a
+          href="/how-it-works"
+          className={`self-stretch my-auto ${
+            isActive("/how-it-works") ? "text-teal-700" : "text-slate-950"
+          }`}
+        >
           How it Works
         </a>
-        <a href="#" className="self-stretch my-auto">
-          Free Resources
+        <a
+          href="#"
+          className={`self-stretch my-auto ${
+            isActive("#") ? "text-teal-700" : "text-slate-950"
+          }`}
+        >
+          What`s New
         </a>
-        <a href="/case-studies" className="self-stretch my-auto">
+        <a
+          href="/case-studies"
+          className={`self-stretch my-auto ${
+            isActive("/case-studies") ? "text-teal-700" : "text-slate-950"
+          }`}
+        >
           Case Studies
         </a>
-        <a href="/about-us" className="self-stretch my-auto">
+        <a
+          href="/about-us"
+          className={`self-stretch my-auto ${
+            isActive("/about-us") ? "text-teal-700" : "text-slate-950"
+          }`}
+        >
           About Us
         </a>
       </nav>
